Re-structure the repo and added poetry related files
