+++ .gitignore
@@ -1,7 +1,8 @@
 .env
 .local/
 poetry.lock
-pyproject.toml
 *.pyc
 __pycache__/
 _test_runner_.py
+.venv/
+.vscode/
