--- checklist.txt
@@ -1,18 +1,0 @@
-History
-  When
-  Who
-Markdown
-  Bold
-  Italic
-  Font Size
-  Images
-  Hyperlinks
-Design
-  Home
-    Search
-    Recent
-    List
-    unlisted entry mode (admins only)
-  Alphabetical
-  Articles
-Pokémon Mode
