--- requirements.txt
@@ -1,9 +1,0 @@
-pytest
-discord
-python-dotenv
-beautifulsoup4
-timeago
-jellyfish
-motor
-dnspython
-pytest-asyncio
