+++ src/config/editors.txt
