+++ src/website/static/markdownshortcuts.js
