+++ src/website/static/quill.css
